♻️ Refactor tests and update PluginId validation logic
Split unit tests into separate files for PluginId, SemanticVersion, and VersionRange. Updated PluginId to use Guard.NotNullOrWhiteSpace for validation. Adjusted project references to remove circular dependency and added Twc.Cli.Shared as a dependency for Twc.Cli.Sdk.
