+++ src/Twc.Cli.Sdk/Twc.Cli.Sdk.csproj
@@ -4,5 +4,8 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
-</Project>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Twc.Cli.Shared\Twc.Cli.Shared.csproj" />
+  </ItemGroup>
+</Project>
