+++ src/Twc.Cli.Shared/Twc.Cli.Shared.csproj
@@ -4,9 +4,4 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Twc.Cli.Sdk\Twc.Cli.Sdk.csproj" />
-  </ItemGroup>
 </Project>
-
